requirements.txt
torch
gradio
transformers
Pillow
pytest
